+++ src/main/java/tech/araopj/jokes/submit/controller/JokeRequestController.java
@@ -1,10 +1,10 @@
-package tech.araopj.jokes.request.controller;
+package tech.araopj.jokes.submit.controller;
 
 import org.springframework.web.server.ResponseStatusException;
-import tech.araopj.jokes.request.service.JokeRequestService;
+import tech.araopj.jokes.submit.service.JokeRequestService;
 import org.springframework.validation.annotation.Validated;
-import tech.araopj.jokes.request.entity.JokeRequestBody;
-import tech.araopj.jokes.request.entity.JokeRequest;
+import tech.araopj.jokes.submit.entity.JokeRequestBody;
+import tech.araopj.jokes.submit.entity.JokeRequest;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
@@ -12,7 +12,7 @@ import static java.lang.String.format;
 import java.util.Collection;
 
 @RestController
-@RequestMapping("v1/request")
+@RequestMapping("v1/submit")
 public record JokeRequestController(JokeRequestService jokeRequestService) {
 
     @GetMapping("/all")
@@ -20,7 +20,7 @@ public record JokeRequestController(JokeRequestService jokeRequestService) {
         return jokeRequestService.getAllJokeRequests();
     }
 
-    @PostMapping("/add")
+    @PostMapping
     public HttpEntity<String> addJoke(@Validated @RequestBody JokeRequestBody jokeRequestBody) throws ResponseStatusException {
         jokeRequestService
                 .doesJokeAlreadySubmitted(jokeRequestBody.joke())
